+++ backend/canisters/user_index/impl/src/updates/delete_user.rs
@@ -1,4 +1,4 @@
-use crate::{mutate_state, DeletedUser, RuntimeState};
+use crate::{mutate_state, RuntimeState};
 use canister_tracing_macros::trace;
 use ic_cdk::update;
 use local_user_index_canister::{DeleteUser, Event};
@@ -20,13 +20,7 @@ fn delete_user_impl(args: Args, state: &mut RuntimeState) -> Response {
         return NotAuthorized;
     }
 
-    let now = state.env.now();
-    state.data.users.delete_user(args.user_id, now);
-    state.data.deleted_users.push(DeletedUser {
-        user_id: args.user_id,
-        triggered_by_user: true,
-        timestamp: now,
-    });
+    state.delete_user(args.user_id, true);
     state.push_event_to_all_local_user_indexes(
         Event::DeleteUser(DeleteUser {
             user_id: args.user_id,