+++ backend/canisters/identity/api/src/updates/mod.rs
@@ -1,5 +1,5 @@
 pub mod approve_identity_link;
-pub mod c2c_sync_legacy_user_principals;
+pub mod c2c_set_user_ids;
 pub mod create_identity;
 pub mod generate_challenge;
 pub mod initiate_identity_link;