+++ backend/canisters/identity/impl/src/updates/mod.rs
@@ -1,4 +1,5 @@
 pub mod approve_identity_link;
+pub mod c2c_set_user_ids;
 pub mod create_identity;
 pub mod generate_challenge;
 pub mod initiate_identity_link;