+++ backend/canisters/user_index/impl/src/lib.rs
@@ -6,7 +6,7 @@ use crate::timer_job_types::TimerJob;
 use candid::Principal;
 use canister_state_macros::canister_state;
 use canister_timer_jobs::TimerJobs;
-use event_store_producer::{EventStoreClient, EventStoreClientBuilder, EventStoreClientInfo};
+use event_store_producer::{EventBuilder, EventStoreClient, EventStoreClientBuilder, EventStoreClientInfo};
 use event_store_producer_cdk_runtime::CdkRuntime;
 use fire_and_forget_handler::FireAndForgetHandler;
 use icrc_ledger_types::icrc1::account::{Account, Subaccount};
@@ -23,7 +23,7 @@ use nns_governance_canister::types::{Empty, ManageNeuron, NeuronId};
 use p256_key_pair::P256KeyPair;
 use serde::{Deserialize, Serialize};
 use std::cell::RefCell;
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::collections::{BTreeMap, HashMap, HashSet, VecDeque};
 use std::time::Duration;
 use types::{
     BuildVersion, CanisterId, CanisterWasm, ChatId, Cryptocurrency, Cycles, DiamondMembershipFees, Milliseconds,
@@ -158,6 +158,36 @@ impl RuntimeState {
         jobs::submit_message_to_modclub::start_job_if_required(self);
     }
 
+    pub fn delete_user(&mut self, user_id: UserId, triggered_by_user: bool) {
+        let now = self.env.now();
+        if let Some(user) = self.data.users.delete_user(user_id, now) {
+            self.data.local_index_map.remove_user(&user_id);
+            self.data.empty_users.remove(&user_id);
+
+            #[derive(Serialize)]
+            struct EventPayload {
+                triggered_by_user: bool,
+            }
+
+            self.data.event_store_client.push(
+                EventBuilder::new("user_deleted", now)
+                    .with_user(user_id.to_string(), true)
+                    .with_source(self.env.canister_id().to_string(), false)
+                    .with_json_payload(&EventPayload { triggered_by_user })
+                    .build(),
+            );
+
+            self.data.deleted_users.push(DeletedUser {
+                user_id,
+                triggered_by_user,
+                timestamp: now,
+            });
+
+            self.data.identity_canister_user_sync_queue.push_back((user.principal, None));
+            jobs::sync_users_to_identity_canister::try_run_now(self);
+        }
+    }
+
     pub fn user_metrics(&self, user_id: UserId) -> Option<UserMetrics> {
         self.data.users.get_by_user_id(&user_id).map(|user| {
             let now = self.env.now();
@@ -233,6 +263,7 @@ impl RuntimeState {
             empty_users_length: self.data.empty_users.len(),
             deleted_users: self.data.deleted_users.iter().take(100).map(|u| u.user_id).collect(),
             deleted_users_length: self.data.deleted_users.len(),
+            unique_person_proofs_submitted: self.data.users.unique_person_proofs_submitted(),
         }
     }
 }
@@ -283,6 +314,8 @@ struct Data {
     pub deleted_users: Vec<DeletedUser>,
     #[serde(default)]
     pub ic_root_key: Vec<u8>,
+    #[serde(default)]
+    pub identity_canister_user_sync_queue: VecDeque<(Principal, Option<UserId>)>,
 }
 
 impl Data {
@@ -352,6 +385,7 @@ impl Data {
             chit_leaderboard: ChitLeaderboard::default(),
             deleted_users: Vec::new(),
             ic_root_key,
+            identity_canister_user_sync_queue: VecDeque::new(),
         };
 
         // Register the ProposalsBot
@@ -459,6 +493,7 @@ impl Default for Data {
             chit_leaderboard: ChitLeaderboard::default(),
             deleted_users: Vec::new(),
             ic_root_key: Vec::new(),
+            identity_canister_user_sync_queue: VecDeque::new(),
         }
     }
 }
@@ -499,6 +534,7 @@ pub struct Metrics {
     pub empty_users_length: usize,
     pub deleted_users: Vec<UserId>,
     pub deleted_users_length: usize,
+    pub unique_person_proofs_submitted: u32,
 }
 
 #[derive(Serialize, Debug)]