+++ backend/canisters/user_index/impl/src/lifecycle/post_upgrade.rs
@@ -4,7 +4,6 @@ use crate::{mutate_state, Data};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk::post_upgrade;
-use local_user_index_canister::{DeleteUser, Event};
 use stable_memory::get_reader;
 use tracing::info;
 use user_index_canister::post_upgrade::Args;
@@ -27,17 +26,12 @@ fn post_upgrade(args: Args) {
     info!(version = %args.wasm_version, "Post-upgrade complete");
 
     mutate_state(|state| {
-        for user_id in std::mem::take(&mut state.data.empty_users) {
-            if let Some(canister_id) = state.data.local_index_map.get_index_canister(&user_id) {
-                state.data.user_index_event_sync_queue.push(
-                    canister_id,
-                    Event::DeleteUser(DeleteUser {
-                        user_id,
-                        triggered_by_user: false,
-                    }),
-                );
-            }
+        for user in state.data.users.iter() {
+            state
+                .data
+                .identity_canister_user_sync_queue
+                .push_back((user.principal, Some(user.user_id)));
         }
-        crate::jobs::sync_events_to_local_user_index_canisters::start_job_if_required(state);
+        crate::jobs::sync_users_to_identity_canister::start_job_if_required(state);
     })
 }