+++ backend/canisters/user_index/CHANGELOG.md
@@ -8,8 +8,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Sync userIds to Identity canister ([#6027](https://github.com/open-chat-labs/open-chat/pull/6027))
+
+## [[2.0.1235](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1235-user_index)] - 2024-07-11
+
+### Added
+
 - Uninstall canisters of empty users ([#6018](https://github.com/open-chat-labs/open-chat/pull/6018))
 - Add `submit_proof_of_unique_personhood` ([#6023](https://github.com/open-chat-labs/open-chat/pull/6023))
+
+### Changed
+
+- Track event each time a proof of uniqueness is submitted ([#6024](https://github.com/open-chat-labs/open-chat/pull/6024))
+- Track event each time a user is deleted ([#6025](https://github.com/open-chat-labs/open-chat/pull/6025))
 
 ## [[2.0.1231](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1231-user_index)] - 2024-07-08